Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -5,6 +5,6 @@ int main()
     char inputCharacter;
     printf("Enter a character: ");
     scanf(" %c", &inputCharacter);
-    printf("ASCII Value: %d\n", inputCharacter);
+    printf("ASCII Value: %d\n", (int)inputCharacter);
     return 0;
 }
